[halfNormal-major-2]
1. major refactoring
