+++ src/pymultifit/distributions/beta_d.py
@@ -3,9 +3,10 @@
 from typing import Dict
 
 import numpy as np
-from scipy.special import beta as beta_function, betainc
+from scipy.special import betainc
 
 from .backend import BaseDistribution
+from .utilities import beta_
 
 
 class BetaDistribution(BaseDistribution):
@@ -19,7 +20,7 @@ class BetaDistribution(BaseDistribution):
         self.norm = normalize
 
     def _pdf(self, x: np.ndarray) -> np.ndarray:
-        return _beta(x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, normalize=self.norm)
+        return beta_(x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, normalize=self.norm)
 
     def pdf(self, x: np.ndarray) -> np.ndarray:
         return self._pdf(x)
@@ -35,38 +36,3 @@ class BetaDistribution(BaseDistribution):
 
         return {'mean': mean_,
                 'variance': variance_}
-
-
-def _beta(x: np.ndarray,
-          amplitude: float = 1., alpha: float = 1., beta: float = 1.,
-          normalize: bool = False) -> np.ndarray:
-    """
-    Compute the beta probability density function (PDF).
-
-    Parameters
-    ----------
-    x : np.ndarray
-        The input array for which to compute the PDF.
-    amplitude : float
-        The amplitude to apply to the PDF. Default is 1.
-    alpha : float
-        The alpha (shape) parameter of the beta distribution. Default is 1.
-    beta : float
-        The beta (shape) parameter of the beta distribution. Default is 1.
-    normalize : bool
-        If True, the PDF is normalized using the beta function. Default is True.
-
-    Returns
-    -------
-    np.ndarray
-        The probability density function values for the given input.
-    """
-    numerator = x**(alpha - 1) * (1 - x)**(beta - 1)
-
-    if normalize:
-        normalization_factor = beta_function(alpha, beta)
-        amplitude = 1.0
-    else:
-        normalization_factor = 1.0
-
-    return amplitude * (numerator / normalization_factor)
