+++ src/pymultifit/distributions/foldedHalfNormal_d.py
@@ -6,7 +6,8 @@ import numpy as np
 from scipy.special import erf
 
 from .backend import BaseDistribution
-from .gaussian_d import gaussian_, GaussianDistribution
+from .gaussian_d import GaussianDistribution
+from .utilities import folded_half_normal_
 
 
 class FoldedHalfNormalDistribution(BaseDistribution):
@@ -48,49 +49,3 @@ class FoldedHalfNormalDistribution(BaseDistribution):
 
         return {'mean': mu_y,
                 'variance': var_y}
-
-
-def folded_half_normal_(x: np.ndarray,
-                        amplitude: float = 1., mu: float = 0.0, variance: float = 1.0,
-                        normalize: bool = False) -> np.ndarray:
-    """
-    Compute the folded half-normal distribution.
-
-    The folded half-normal distribution is the sum of two Gaussian distributions mirrored around `mu`.
-    It is defined as the sum of a normal distribution centered at `mu` and another mirrored at `-mu`.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Input array where the folded half-normal distribution is evaluated.
-    amplitude : float, optional
-        The amplitude of the distribution. Defaults to 1. Ignored if `normalize` is set to True.
-    mu : float, optional
-        The mean (`mu`) of the original normal distribution. Defaults to 0.0.
-    variance : float, optional
-        The standard deviation (`sigma`) of the original normal distribution. Defaults to 1.0.
-    normalize : bool, optional
-        If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
-
-    Returns
-    -------
-    np.ndarray
-        The computed folded half-normal distribution values for the input `x`.
-
-    Notes
-    -----
-    The folded half-normal distribution is defined as the sum of two Gaussian
-    distributions:
-    .. math::
-        f(x; \\mu, \\sigma) = g_1(x; \\mu, \\sigma) + g_2(x; -\\mu, \\sigma)
-
-    where `g_1` and `g_2` are the Gaussian distributions with the specified parameters.
-    """
-    sigma = np.sqrt(variance)
-    mask = x >= 0
-    g1 = gaussian_(x[mask], amplitude=amplitude, mu=mu, sigma=sigma, normalize=normalize)
-    g2 = gaussian_(x[mask], amplitude=amplitude, mu=-mu, sigma=sigma, normalize=normalize)
-    result = np.zeros_like(x)
-    result[mask] = g1 + g2
-
-    return result
