+++ src/pymultifit/distributions/powerLaw_d.py
@@ -5,6 +5,7 @@ from typing import Any, Dict
 import numpy as np
 
 from .backend import BaseDistribution
+from .utilities import power_law_
 
 
 class PowerLawDistribution(BaseDistribution):
@@ -26,29 +27,3 @@ class PowerLawDistribution(BaseDistribution):
 
     def stats(self) -> Dict[str, Any]:
         pass
-
-
-def power_law_(x: np.ndarray, amplitude: float = 1, alpha: float = -1, normalize: bool = False) -> np.ndarray:
-    """
-    Compute a power-law function for a given set of x values.
-
-    This function is designed with a `normalize` parameter for consistency with other probability density functions (PDFs).
-    However, the `normalize` parameter has no effect in this function, as normalization of the power law is not handled here.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Input values for which the power-law function will be evaluated.
-    amplitude : float, optional
-        The amplitude or scaling factor of the power law, by default 1.
-    alpha : float, optional
-        The exponent of the power law, by default -1.
-    normalize : bool, optional
-        Included for consistency with other PDF functions. Has no effect on the output. Defaults to False.
-
-    Returns
-    -------
-    np.ndarray
-        Computed power-law values for each element in x.
-    """
-    return amplitude * x**-alpha
