+++ src/pymultifit/fitters/gaussian_f.py
@@ -1,14 +1,15 @@
 """Created on Jul 18 00:25:57 2024"""
 
-from .backend import BaseFitter, utilities as utils
-from ..distributions import gaussian_
+from .backend import BaseFitter
+from .utilities import sanity_check
+from ..distributions.utilities import gaussian_
 
 
 class GaussianFitter(BaseFitter):
     """A class for fitting multiple Gaussian functions to the given data."""
 
     def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = utils.sanity_check(x_values=x_values, y_values=y_values)
+        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(n_fits=n_fits, x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 3
