+++ src/pymultifit/fitters/logNorm_f.py
@@ -1,7 +1,8 @@
 """Created on Jul 18 19:01:45 2024"""
 
-from .backend import BaseFitter, utilities as utils
-from ..distributions import log_normal_
+from .backend import BaseFitter
+from .utilities import sanity_check
+from ..distributions.utilities import log_normal_
 
 
 # TODO:
@@ -12,7 +13,7 @@ class LogNormalFitter(BaseFitter):
     """A class for fitting multiple Log Normal functions to the given data."""
 
     def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = utils.sanity_check(x_values=x_values, y_values=y_values)
+        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         if any(x_values < 0):
             raise ValueError("The LogNormal distribution must have x > 0. "
                              "Use `EPSILON` from the package to get as close to zero as possible.")
