+++ src/pymultifit/fitters/norris_f.py
@@ -1,14 +1,15 @@
 """Created on Nov 30 05:33:38 2024"""
 
-from .backend import BaseFitter, utilities as utils
-from ..distributions import norris2005, norris2011
+from .backend import BaseFitter
+from .utilities import sanity_check
+from ..distributions.utilities import norris2005, norris2011
 
 
 class Norris2005Fitter(BaseFitter):
     """A class for fitting multiple Norris 2005 models to the given data."""
 
     def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = utils.sanity_check(x_values=x_values, y_values=y_values)
+        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(n_fits=n_fits, x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 3
 
@@ -21,7 +22,7 @@ class Norris2011Fitter(BaseFitter):
     """A class for fitting multiple Norris 2011 models to the given data."""
 
     def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = utils.sanity_check(x_values=x_values, y_values=y_values)
+        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(n_fits=n_fits, x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 3
