+++ src/pymultifit/fitters/skewNorm_f.py
@@ -2,14 +2,15 @@
 
 from scipy.stats import skewnorm
 
-from .backend import BaseFitter, utilities as utils
+from .backend import BaseFitter
+from .utilities import sanity_check
 
 
 class SkewedNormalFitter(BaseFitter):
     """A class for fitting multiple Skewed Normal functions to the given data."""
 
     def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = utils.sanity_check(x_values=x_values, y_values=y_values)
+        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(n_fits=n_fits, x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 4
